Skrive tester for å hente inn arbeidsforhold på person + flytte tester
